Bugfix: support for # in topic does not dispatch correctly
Subscribing on wildcard topics works, but upon dispatch the topic handler
cannot be found.
This patch parses the topicMap for topics ending in '#' in case no exact
match exists. This means that:

- the topic handler subscribed to the exact topic is called
- if no match was found, a matching wildcard handler is called
- if still no match was found, the topic is ignored

(in case of multiple overlapping subscriptions such as  'mytopic/#'
and 'mytopic/subtopic/#' only one of the handlers is called, this is not
really intended to be supported)